Focus the HTML text editor as soon as it's loaded.
Note: This works ok for us since we're always hiding the editor behind a
placeholder element and then we show it once clicked on the placeholder
element. Though, this logic might be undesirable if we ever want to show
the editor without the placeholder element, in which case it will be
focused as soon as the page is opened.