Implement some common table operations.
It is now possible to add a new row before/after the one that's
currently focused, as well as add a new column left/right of the one
that's currently focused. You can also delete a whole row and a whole
column, in which case a "previous" row/column will become focused next.