Add the `versionedAt` property to the published repository info
The `versionedAt` property represents a date which is updated whenever
something gets published inside a specific repository. It is a
compliment to the `version` property, which is a hash representation of
the same thing.

Another difference is that the `versionedAt` property is included inside
the repository catalog info as well.