Remove the always showing scrollbar on an html element set by Vuetify
So apparently a part of the Vuetify's (questionable) design is to set
`overflow-y: scroll` on an `html` element, which makes the scrollbar
visible even when there is no overflowing content.
See this issue for more info: https://github.com/vuetifyjs/vuetify/issues/864

Note: Mac users probably don't notice this issue, because MacOS hides
the scrollbar by default if there's no mouse connected to it.