Stop linting the extensions folder
Since the extensions folder should, by definition, contain the
third-party code, it doesn't make much sense to run the linter on it.
Especially since the build will fail if there are any linter errors,
which most certainly there will be if the extensions don't follow
exactly the same eslint rules as the tailor itself does.