Update dependencies
Almost all dependencies are updated to the latest version.
Exceptions are dependencies related to Vue, as migration to v3
is a separate task. Dependencies blocked by other dependencies
based on the required peer version, `lerna` as the latest version
removed deprecated commands used by the app, and `yup` as there is
a bug in the latest version which breaks the app.