Update `packages` related dependencies
The list of internal dependencies from the `packages` folder,
`@tailor-cms/config`, `@tailor-cms/core-components`, `@tailor-cms/utils`,
and `@extensionengine/vue-radio` are updated to the latest version.
The `vuetify` dependency is updated to satisfy the latest peer dependency
definition from the `@tailor-cms/core-components`.
The `@extensionengine/stylelint-config` and `stylelint` dependencies are
updated to resolve conflicting versions of `stylelint` at the root level.